+++ StswExpress.Commons/Utils/Collections/_items/StswComboItem.cs
@@ -8,10 +8,20 @@ public partial class StswComboItem : StswObservableObject
     /// <summary>
     /// Gets or sets the display text for the item.
     /// </summary>
-    [StswObservableProperty] object? _display;
+    public object? Display
+    {
+        get => _display;
+        set => SetProperty(ref _display, value);
+    }
+    private object? _display;
 
     /// <summary>
     /// Gets or sets the value associated with the item.
     /// </summary>
-    [StswObservableProperty] object? _value;
+    public object? Value
+    {
+        get => _value;
+        set => SetProperty(ref _value, value);
+    }
+    private object? _value;
 }