+++ StswExpress.Commons/Utils/Mailboxes/StswMailboxModel.cs
@@ -17,9 +17,11 @@ public partial class StswMailboxModel : StswObservableObject
     */
 
     public StswMailboxModel() { }
-    public StswMailboxModel(string? host = null, string? from = null, string? username = null, string? password = null)
+    public StswMailboxModel(string? host = null, string? from = null, string? username = null, string? password = null) : this(host, null, from, username, password) { }
+    public StswMailboxModel(string? host = null, int? port = null, string? from = null, string? username = null, string? password = null)
     {
         Host = host;
+        Port = port;
         From = from;
         Username = username;
         Password = password;
@@ -28,47 +30,106 @@ public partial class StswMailboxModel : StswObservableObject
     /// <summary>
     /// Gets or sets the name of the email account.
     /// </summary>
-    [StswObservableProperty] string? _name;
+    public string? Name
+    {
+        get => _name;
+        set => SetProperty(ref _name, value);
+    }
+    private string? _name;
 
     /// <summary>
     /// Gets or sets the SMTP host for the email account.
     /// </summary>
-    [StswObservableProperty] string? _host;
+    public string? Host
+    {
+        get => _host;
+        set => SetProperty(ref _host, value);
+    }
+    private string? _host;
 
     /// <summary>
     /// Gets or sets the port number for the SMTP host.
     /// </summary>
-    [StswObservableProperty] int? _port = 587;
+    public int? Port
+    {
+        get => _port;
+        set => SetProperty(ref _port, value);
+    }
+    private int? _port = 587;
 
     /// <summary>
     /// Gets or sets the email address associated with the email account.
     /// </summary>
-    [StswObservableProperty] string? _from;
+    public string? From
+    {
+        get => _from;
+        set => SetProperty(ref _from, value);
+    }
+    private string? _from;
 
     /// <summary>
     /// Gets or sets the username for the email account.
     /// </summary>
-    [StswObservableProperty] string? _username;
+    public string? Username
+    {
+        get => _username;
+        set => SetProperty(ref _username, value);
+    }
+    private string? _username;
 
     /// <summary>
     /// Gets or sets the password for the email account.
     /// </summary>
-    [StswObservableProperty] string? _password;
+    public string? Password
+    {
+        get => _password;
+        set => SetProperty(ref _password, value);
+    }
+    private string? _password;
 
     /// <summary>
     /// Gets or sets the domain for the email account.
     /// </summary>
-    [StswObservableProperty] string? _domain;
+    [StswInfo("0.8.2")]
+    public string? Domain
+    {
+        get => _domain;
+        set => SetProperty(ref _domain, value);
+    }
+    private string? _domain;
 
     /// <summary>
     /// Gets or sets the collection of reply-to addresses for the email account.
     /// </summary>
-    [StswObservableProperty] IEnumerable<string>? _replyTo;
+    [StswInfo("0.9.0")]
+    public IEnumerable<string>? ReplyTo
+    {
+        get => _replyTo;
+        set => SetProperty(ref _replyTo, value);
+    }
+    private IEnumerable<string>? _replyTo;
+
+    /// <summary>
+    /// Gets or sets a value indicating whether to ignore SSL certificate errors when connecting to the SMTP server.
+    /// </summary>
+    [StswInfo("0.19.1")]
+    public bool IgnoreCertificateErrors
+    {
+        get => _ignoreCertificateErrors;
+        set => SetProperty(ref _ignoreCertificateErrors, value);
+    }
+    private bool _ignoreCertificateErrors;
 
     /// <summary>
     /// Gets or sets the security option for the SMTP connection.
     /// </summary>
-    [StswObservableProperty] SecureSocketOptions _securityOption = SecureSocketOptions.StartTls;
+    [StswInfo("0.19.0")]
+    public SecureSocketOptions SecurityOption
+    {
+        get => _securityOption;
+        set => SetProperty(ref _securityOption, value);
+    }
+    private SecureSocketOptions _securityOption = SecureSocketOptions.Auto;
 
     /// <summary>
     /// Sends an email using the SMTP protocol with optional attachments, BCC recipients, and reply-to addresses.
@@ -79,9 +140,30 @@ public partial class StswMailboxModel : StswObservableObject
     /// <param name="body">The body content of the email.</param>
     /// <param name="attachments">An optional collection of file paths to attach to the email.</param>
     /// <param name="bcc">An optional collection of BCC recipients.</param>
-    /// <param name="reply">An optional collection of reply-to addresses.</param>
     public void Send(IEnumerable<string> to, string subject, string body, bool? isBodyHtml = null, IEnumerable<string>? attachments = null, IEnumerable<string>? cc = null, IEnumerable<string>? bcc = null)
-        => SendAsync(to, subject, body, isBodyHtml, attachments, cc, bcc).GetAwaiter().GetResult();
+    {
+        if (!StswMailboxes.Config.IsEnabled)
+            return;
+
+        // if (!CanSendEmail())
+        //     return;
+
+        ArgumentException.ThrowIfNullOrEmpty(Host);
+        ArgumentException.ThrowIfNullOrEmpty(From);
+        ArgumentNullException.ThrowIfNull(Port);
+
+        var message = BuildMessage(to, subject, body, isBodyHtml ?? false, attachments, cc, bcc);
+
+        using var client = CreateConfiguredClient();
+        client.Connect(Host, Port.Value, SecurityOption);
+        if (!string.IsNullOrEmpty(Username))
+            client.Authenticate(Username, Password);
+
+        client.Send(message);
+        client.Disconnect(true);
+
+        // IncrementEmailCounters();
+    }
 
     /// <summary>
     /// Sends an email using the SMTP protocol with optional attachments.
@@ -111,7 +193,6 @@ public partial class StswMailboxModel : StswObservableObject
     /// <param name="body">The body content of the email.</param>
     /// <param name="attachments">An optional collection of file paths to attach to the email.</param>
     /// <param name="bcc">An optional collection of BCC recipients.</param>
-    /// <param name="reply">An optional collection of reply-to addresses.</param>
     public async Task SendAsync(IEnumerable<string> to, string subject, string body, bool? isBodyHtml = null, IEnumerable<string>? attachments = null, IEnumerable<string>? cc = null, IEnumerable<string>? bcc = null)
     {
         if (!StswMailboxes.Config.IsEnabled)
@@ -124,37 +205,10 @@ public partial class StswMailboxModel : StswObservableObject
         ArgumentException.ThrowIfNullOrEmpty(From);
         ArgumentNullException.ThrowIfNull(Port);
 
-        var message = new MimeMessage();
-        message.From.Add(MailboxAddress.Parse(From));
+        var message = BuildMessage(to, subject, body, isBodyHtml ?? false, attachments, cc, bcc);
 
-        if (!string.IsNullOrEmpty(StswMailboxes.Config.DebugEmailRecipient) && StswFn.IsInDebug())
-        {
-            message.To.Add(MailboxAddress.Parse(StswMailboxes.Config.DebugEmailRecipient));
-        }
-        else
-        {
-            to?.ForEach(x => message.To.Add(MailboxAddress.Parse(x)));
-            cc?.ForEach(x => message.Cc.Add(MailboxAddress.Parse(x)));
-            bcc?.ForEach(x => message.Bcc.Add(MailboxAddress.Parse(x)));
-            ReplyTo?.ForEach(x => message.ReplyTo.Add(MailboxAddress.Parse(x)));
-        }
-        
-        message.Subject = subject;
-        var builder = new BodyBuilder
-        {
-            HtmlBody = (isBodyHtml ?? false) ? body : null,
-            TextBody = (isBodyHtml ?? false) ? null : body
-        };
-
-        if (attachments != null)
-            foreach (var file in attachments)
-                builder.Attachments.Add(file);
-
-        message.Body = builder.ToMessageBody();
-
-        using var client = new SmtpClient();
+        using var client = CreateConfiguredClient();
         await client.ConnectAsync(Host, Port.Value, SecurityOption);
-        
         if (!string.IsNullOrEmpty(Username))
             await client.AuthenticateAsync(Username, Password);
 
@@ -182,6 +236,64 @@ public partial class StswMailboxModel : StswObservableObject
     /// <param name="body">The body content of the email.</param>
     public Task SendAsync(IEnumerable<string> to, string subject, string body)
         => SendAsync(to, subject, body, null, null, null);
+
+    /// <summary>
+    /// Builds a <see cref="MimeMessage"/> for sending an email with optional attachments, BCC recipients, and reply-to addresses.
+    /// </summary>
+    /// <param name="to">The collection of recipient email addresses.</param>
+    /// <param name="subject">The subject of the email.</param>
+    /// <param name="body">The body content of the email.</param>
+    /// <param name="isBodyHtml">Indicates whether the body content is in HTML format.</param>
+    /// <param name="attachments">An optional collection of file paths to attach to the email.</param>
+    /// <param name="cc">The collection of CC recipients.</param>
+    /// <param name="bcc">The collection of BCC recipients.</param>
+    /// <returns>A configured <see cref="MimeMessage"/> instance.</returns>
+    [StswInfo("0.19.1")]
+    private MimeMessage BuildMessage(IEnumerable<string> to, string subject, string body, bool isBodyHtml, IEnumerable<string>? attachments = null, IEnumerable<string>? cc = null, IEnumerable<string>? bcc = null)
+    {
+        var message = new MimeMessage();
+        message.From.Add(MailboxAddress.Parse(From));
+
+        if (!string.IsNullOrEmpty(StswMailboxes.Config.DebugEmailRecipient) && StswFn.IsInDebug())
+        {
+            message.To.Add(MailboxAddress.Parse(StswMailboxes.Config.DebugEmailRecipient));
+        }
+        else
+        {
+            to?.ForEach(x => message.To.Add(MailboxAddress.Parse(x)));
+            cc?.ForEach(x => message.Cc.Add(MailboxAddress.Parse(x)));
+            bcc?.ForEach(x => message.Bcc.Add(MailboxAddress.Parse(x)));
+            ReplyTo?.ForEach(x => message.ReplyTo.Add(MailboxAddress.Parse(x)));
+        }
+
+        message.Subject = subject;
+
+        var builder = new BodyBuilder
+        {
+            HtmlBody = isBodyHtml ? body : null,
+            TextBody = isBodyHtml ? null : body
+        };
+
+        attachments?.ForEach(file => builder.Attachments.Add(file));
+        message.Body = builder.ToMessageBody();
+
+        return message;
+    }
+
+    /// <summary>
+    /// Creates a configured <see cref="SmtpClient"/> instance for sending emails.
+    /// </summary>
+    /// <returns>A configured <see cref="SmtpClient"/> instance.</returns>
+    [StswInfo("0.19.1")]
+    private SmtpClient CreateConfiguredClient()
+    {
+        var client = new SmtpClient();
+
+        if (IgnoreCertificateErrors)
+            client.ServerCertificateValidationCallback = (s, c, chain, sslPolicyErrors) => true;
+
+        return client;
+    }
 
     /*
     private bool CanSendEmail()